// packages/helpers/src/getEnv/getEnv.types.ts
export enum DS_ENV {
  BASE_BACKGROUND = "BASE_BACKGROUND",
  BASE_ACCENT_1 = "BASE_ACCENT_1",
  BASE_ACCENT_2 = "BASE_ACCENT_2",
  BASE_ACCENT_3 = "BASE_ACCENT_3",
  BASE_FOREGROUND = "BASE_FOREGROUND",
  BASE_SUCCESS = "BASE_SUCCESS",
  BASE_ERROR = "BASE_ERROR",
  BASE_ALERT = "BASE_ALERT",
  BASE_INFO = "BASE_INFO",
}

export enum WEB_ENV {
  API_BASE_URL = "API_BASE_URL",
}

export enum API_ENV {}
// Add api-specific environment variables here

export enum STORYBOOK_ENV {}
// Add storybook-specific environment variables here